HOTFIX
Se quita la implementación de librerías de Spring del modelo porque irrespetaban la arquitectura hexagona.
